Ypdated scripts to remove multiple requests for password

+++ Distribution/HappyBrackets Developer Kit/HappyBrackets Project/config/update-device.sh
@@ -1,10 +1,17 @@
 #!/bin/bash
 
+
+
 ### Script to upload new network to device by automating the scp command
 ### The device is added as a parameter to the script
 
-### move to the correct dir for running this script 
+### move to the correct dir for running this script
 cd ../../Device
+
+
+#we need a copy of this becasue we will need it in the spawn
+CURRENT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+echo "Current dir" $CURRENT_DIR
 
 
 if [ $# -gt 0 ]; then
@@ -17,25 +24,94 @@ else
    
 fi
 
+## first see if we can use expect to save our passwords inside the script
+if [ -f "/usr/bin/expect" ]
+then
+    DEF_PASSWORD="raspberry"
+
+    while true
+    do
+        echo "Enter Password. Just press enter for default "$DEF_PASSWORD
+        read -s PASSWORD
+
+        ENTRY_LEN=${#PASSWORD}
+
+        if [ ${#PASSWORD}  -eq 0 ]
+        then
+            PASSWORD=$DEF_PASSWORD
+            echo "Use Default password"
+            break
+
+        else
+            #we need to check that the same password is used
+            echo "Re Enter Password."
+            read -s PASSWORD_RETRY
+            if [ "$PASSWORD" = "$PASSWORD_RETRY" ]
+            then
+                break
+            else
+                echo "Passwords do not match"
+            fi
+        fi
+    done
+fi
+
+#define a function where we can use our SSH commands in a shell and pass our password in so we don't ned to interact
+function expect_password {
+    expect -c "\
+    set timeout 90
+    set env(TERM)
+    spawn $1
+    expect \"assword:\"
+    send \"$PASSWORD\r\"
+    expect eof
+  "
+}
 
 if [ "$DEVICE_NAME" != "" ]; then
 # run scp 
     echo "Update device"
 
+    echo $PASSWORD
     while true
     do
 
         HOST_ADDRESS="pi@${DEVICE_NAME}"
         echo “Running Upload to ${HOST_ADDRESS}”
-        scp -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -r HappyBrackets/* $HOST_ADDRESS:~/HappyBrackets
 
-        # now we need to SSH into device so we can do a restart of PI
-        echo "We need to reboot our device"
+        ## See if we can automate this with our stored password
+        if [ -f "/usr/bin/expect" ]
+        then
 
-        echo "sudo shutdown -r now"
+            FILES_TO_SEND="$CURRENT_DIR"/HappyBrackets/*
+            FILES_TO_SEND="HappyBrackets/*"
 
-        ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no $HOST_ADDRESS "sudo shutdown -r now"
+            ls $FILES_TO_SEND
 
+            echo "Sending"$FILES_TO_SEND
+
+            expect_password "sh -c \"scp -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -r HappyBrackets/* $HOST_ADDRESS:~/HappyBrackets\""
+
+            # now we need to SSH into device so we can do a restart of PI
+            echo "We need to reboot our device"
+
+            echo "sudo shutdown -r now"
+
+            #Do a spawn to reboot
+
+            expect_password "ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no $HOST_ADDRESS \"sudo shutdown -r now\""
+
+
+        else
+            scp -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -r HappyBrackets/* $HOST_ADDRESS:~/HappyBrackets
+
+            # now we need to SSH into device so we can do a restart of PI
+            echo "We need to reboot our device"
+
+            echo "sudo shutdown -r now"
+
+            ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no $HOST_ADDRESS "sudo shutdown -r now"
+        fi
 	#see if we want to update another device
         while true
         do
