+++ Distribution/HappyBrackets Developer Kit/HappyBrackets Project/data/send-jars.sh
@@ -13,11 +13,58 @@ else
    DEVICE_NAME=$DEVICE
 fi
 
-    echo "Jar files will be sent from this folder"
+echo "Jar files will be sent from this folder"
+
+FILE_PATH="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+echo "Current dir" $FILE_PATH
+
+echo "Send jars FROM ${FILE_PATH} to ${DEVICE_NAME} "
 
 
-    FILE_PATH=$(dirname $0)
-    echo "Send jars FROM ${FILE_PATH} to ${DEVICE_NAME} "
+## first see if we can use expect to save our passwords inside the script
+if [ -f "/usr/bin/expect" ]
+then
+    DEF_PASSWORD="raspberry"
+
+    while true
+    do
+        echo "Enter Password. Just press enter for default "$DEF_PASSWORD
+        read -s PASSWORD
+
+        ENTRY_LEN=${#PASSWORD}
+
+        if [ ${#PASSWORD}  -eq 0 ]
+        then
+            PASSWORD=$DEF_PASSWORD
+            echo "Use Default password"
+            break
+
+        else
+            #we need to check that the same password is used
+            echo "Re Enter Password."
+            read -s PASSWORD_RETRY
+            if [ "$PASSWORD" = "$PASSWORD_RETRY" ]
+            then
+                break
+            else
+                echo "Passwords do not match"
+            fi
+        fi
+    done
+fi
+
+
+#define a function where we can use our SSH commands in a shell and pass our password in so we don't ned to interact
+function expect_password {
+    expect -c "\
+    set timeout 90
+    set env(TERM)
+    spawn $1
+    expect \"assword:\"
+    send \"$PASSWORD\r\"
+    expect eof
+  "
+}
 
 
 
@@ -27,18 +74,38 @@ if [ "$DEVICE_NAME" != "" ]; then
     do
         HOST_ADDRESS="pi@${DEVICE_NAME}"
         echo “Sending jar files to ${HOST_ADDRESS}”
-        scp -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -r $FILE_PATH/jars/* $HOST_ADDRESS:~/HappyBrackets/data/jars/
 
-        # now we need to SSH into device so we can do a restart of PI
-        echo "We need to reboot our device"
+        ## See if we can automate this with our stored password
+        if [ -f "/usr/bin/expect" ]
+        then
 
-        echo "sudo shutdown -r now"
-
-        ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no $HOST_ADDRESS "sudo shutdown -r now"
+            expect_password "sh -c \"scp -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -r jars/* $HOST_ADDRESS:~/HappyBrackets/data/jars/\""
 
 
+            # now we need to SSH into device so we can do a restart of PI
+            echo "We need to reboot our device"
 
-#see if we want to update another device
+            echo "sudo shutdown -r now"
+
+            #Do a spawn to reboot
+
+            expect_password "ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no $HOST_ADDRESS \"sudo shutdown -r now\""
+
+
+        else
+            scp -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -r $FILE_PATH/jars/* $HOST_ADDRESS:~/HappyBrackets/data/jars/
+
+            # now we need to SSH into device so we can do a restart of PI
+            echo "We need to reboot our device"
+
+            echo "sudo shutdown -r now"
+
+            ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no $HOST_ADDRESS "sudo shutdown -r now"
+
+        fi
+
+
+        #see if we want to update another device
         while true
         do
 	    #see if we need to send to more devices
